% Generated by roxygen2 (4.0.0): do not edit by hand
\name{create_ggtheme}
\alias{create_ggtheme}
\title{Create customised themes for ggplot2 object}
\usage{
create_ggtheme(theme = "blank")
}
\arguments{
\item{theme}{The name of the temaplate (blank, xkcd, more to come ...)}
}
\description{
This function creates a ggplot2 theme object based on my favourite templates.
}
\examples{
theme_blank <- create_ggtheme("blank")
theme_xkcd <- create_ggtheme("xkcd")
}

