% Generated by roxygen2 (4.0.1): do not edit by hand
\name{display_colours}
\alias{display_colours}
\title{Display colour using rblocks}
\usage{
display_colours(inp_col)
}
\arguments{
\item{inp_col}{Colours input as names or hex code}
}
\description{
This function display colours using rblocks
}
\examples{
set.seed(1234)
pal_pf <- extract_colours("http://www.scoutlondon.com/blog/wp-content/uploads/2012/05/Pulp-Fiction.jpg")
display_colours(pal_pf)
}

