% Generated by roxygen2 (4.0.0): do not edit by hand
\name{extract_colours}
\alias{extract_colours}
\title{Detect and extract dominant colours from an image}
\usage{
extract_colours(url_img = "http://developer.r-project.org/Logo/Rlogo-1.png",
  num_col = 5, nme_pal = "custom_palette")
}
\arguments{
\item{url_img}{URL of the image (can be PNG, JPG, JPEG, GIF, BMP)}

\item{num_col}{Number of dominant colours to be extracted (default = 5)}

\item{nme_pal}{Name of the new palette (default = "custom_palette")}
}
\description{
This function downloads an image and extracts the dominant colours from it.
The code is based on dsparks' gist (https://gist.github.com/dsparks/3980277).
}
\examples{
## Using the R Logo to create a 5-colour palette
pal_r <- extract_colours("http://developer.r-project.org/Logo/Rlogo-1.png")
par(mfrow = c(1,2))
pie(rep(1, 5), col = pal_r, main = "Palette based on R Logo")
hist(Nile, breaks = 5, col = pal_r, main = "Palette based on R Logo")
}

