% Generated by roxygen2 (4.0.0): do not edit by hand
\name{interpolate_colours}
\alias{interpolate_colours}
\title{Interpolate colour palette to render more colours within the input range}
\usage{
interpolate_colours(inp_col, n_col)
}
\arguments{
\item{inp_col}{Colours input as names or hex code}

\item{n_col}{Number of output colours}
}
\description{
This function extend a colour palette by interpolation.
}
\examples{
set.seed(1234)
inp_col <- extract_colours("http://www.scoutlondon.com/blog/wp-content/uploads/2012/05/Pulp-Fiction.jpg")
out_col <- interpolate_colours(inp_col, n_col = 20)
display_colours(out_col)
}

