% Generated by roxygen2 (4.0.0): do not edit by hand
\name{palette_reduce}
\alias{palette_reduce}
\title{Reduce palette by removing colours with low saturation, or low or high luminance}
\usage{
palette_reduce(pal, sat.thresh = NULL, light.thresh = NULL,
  dark.thresh = NULL)
}
\arguments{
\item{pal}{A palette as a character vector containing hexidecimal coded colours}

\item{sat.thresh}{Minimum saturation of colours in resulting palette (between 0 and 1)}

\item{light.thresh}{Maximum luminosity of colours in resulting palette (between 0 and 1)}

\item{dark.thresh}{Minimum luminosity of colours in resulting palette (between 0 and 1)}
}
\description{
Reduce palette by removing colours with low saturation, or low or high luminance
}

