\name{plot_radar}
\alias{plot_radar}
\title{Create radar (spider) charts with a data frame}
\usage{
plot_radar(data, data.row = NULL, y.cols = NULL, main = NULL, add = F,
  col = "red", lty = 1, scale = T)
}
\description{
Original Code:
http://statisticstoproveanything.blogspot.co.uk/2013/11/spider-web-plots-in-r.html
}

