% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulate_colours}
\alias{simulate_colours}
\title{Simulate Three Different Types of Colour Blindness (Deuteranopia, Protanopia & Tritanopia)}
\usage{
simulate_colours(inp_col)
}
\arguments{
\item{inp_col}{Original Colours (Hex Code or Colour Names)}
}
\description{
This function simulates three different types of colour blindness (http://www.color-blindness.com/2010/03/09/types-of-color-blindness/).
}
\examples{
set.seed(1234)
pal_pf <- extract_colours("http://www.scoutlondon.com/blog/wp-content/uploads/2012/05/Pulp-Fiction.jpg")
simulate_colours(pal_pf)
}

