% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mpd_select_colours}
\alias{mpd_select_colours}
\title{Select colours from a palette to maximize perceptual distance between the colours}
\usage{
mpd_select_colours(pal, sat.thresh = NULL, light.thresh = NULL,
  dark.thresh = NULL, nreps = 10000, ncolours = ceiling(length(pal)/2),
  nreturn = 1)
}
\arguments{
\item{pal}{A palette as a character vector containing hexidecimal coded colours}

\item{sat.thresh}{Minimum saturation of colours in resulting palette (between 0 and 1)}

\item{light.thresh}{Maximum luminosity of colours in resulting palette (between 0 and 1)}

\item{sat.thresh}{Minimum luminosity of colours in resulting palette (between 0 and 1)}

\item{nreps}{The number of samples of the new colour palette to perform for brute force optimization of MPD}
}
\description{
This function takes a palette as a character vector of hexidecimal colours, and returns a smaller palette,
attempting to maximize the mean perceptual distance (MPD) between the colours in the new palette. It uses CIELAB
colorspace to map colour to a perceptually meaningful scale before maximizing distances
}

