% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_radar}
\alias{plot_radar}
\title{Create radar (spider) charts from a data frame}
\usage{
plot_radar(data, data.row = NULL, y.cols = NULL, main = NULL, add = F,
  col = "red", lty = 1, scale = T)
}
\description{
This function creates a radar chart.
Original Code: http://statisticstoproveanything.blogspot.co.uk/2013/11/spider-web-plots-in-r.html
}

